read LG csvs urls from a dict & fetch only if the file does not exist

+++ fetch_raw_data.py
@@ -157,13 +157,19 @@ def fetch_from_recovery_csv(user, google_cloud_credentials_file_path, raw_data_d
         log.info(f"Exported TracedData")
 
 def fetch_listening_groups_csvs(google_cloud_credentials_file_path, pipeline_configuration, raw_data_dir):
-    for listening_group_csv_url in pipeline_configuration.listening_group_csv_urls:
-        listening_group = listening_group_csv_url.split("/")[-1]
+    for k, v in pipeline_configuration.listening_group_csv_urls.items():
+        for i, listening_group_csv_url in enumerate(v):
+            listening_group = listening_group_csv_url.split("/")[-1]
 
-        log.info(f"Saving '{listening_group}' to file '{raw_data_dir}'...")
-        with open(f'{raw_data_dir}/{listening_group}', "wb") as listening_group_output_file:
-            google_cloud_utils.download_blob_to_file(
-                google_cloud_credentials_file_path, listening_group_csv_url, listening_group_output_file)
+            if os.path.exists(f'{raw_data_dir}/{listening_group}'):
+                log.info(f"File '{raw_data_dir}/{listening_group}' for '{listening_group}' already exists;"
+                         f" skipping download")
+                continue
+
+            log.info(f"Saving '{listening_group}' to file '{raw_data_dir}'...")
+            with open(f'{raw_data_dir}/{listening_group}', "wb") as listening_group_output_file:
+                google_cloud_utils.download_blob_to_file(
+                    google_cloud_credentials_file_path, listening_group_csv_url, listening_group_output_file)
 
 def main(user, google_cloud_credentials_file_path, pipeline_configuration_file_path, raw_data_dir):
     # Read the settings from the configuration file
